+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [NEXT_RELEASE]
 ### Added
-- Command `build list`
 - Limits to the nginx container
+- Commands `build list` and `build run`
 
 ### Fixed
 - [server] Fix panic when monitoring the rolling update